Added an example, deep self-organized map
The deep self-organized map (DSOM) is a tool that simultaneously identifies and organizes latent space for the input dataset. Here I apply the DSOM to spectroscopic (graph) data that I preprocessed myself. In the JNB, I use the trained DSOM to investigate connections between latent space and various known characteristics of the galaxies.