Example data visualizations for deep learning
In this notebook, I load in model classifications that I used in a recent deep learning / machine vision project. The project involves a 5x5 matrix of trained CNN models and test datasets. I benchmark the performance of each model on each of the 5 datasets, and assess the scientific impact of their performance stats.